State design pattern / FSM refactor for enemies
Completely refactored the enemy logic to utilize the state design pattern. Now, each state is initialized in the EnemyStateManager and managed entirely by it so that only one state's script can be running at a time and transitions are all managed through a central location.

Also, added multiple fixes for enemy behavior including falling correctly when stunned / disabled and when colliding with an enemy by jumping on their head the player will now be pushed off before starting the QTE if there is space